Compiler: Scratching Flex/Bison for ANTLR 4.7.1
Flex and Bison were becoming a pain to integrate with C++ and C++11. The
lack of documentation was also unhelpful, compounded by the ugly code.
Classmates suggested ANTLR which has plugins for all IDEs and has much
more robust code. It supports C++14 and creates a parse tree for you. I
was able to get this demo working in about an hour. Will build off of
this demo.